user updates when a game is won on game 00
