+++ client/store/user.js
@@ -1,6 +1,7 @@
 import axios from 'axios';
 
 const GET_SINGLE_USER = 'GET_SINGLE_USER';
+const UPDATE_USER = "UPDATE_USER";
 
 const _getSingleUser = (data) => {
   return {
@@ -8,6 +9,11 @@ const _getSingleUser = (data) => {
     data,
   };
 };
+
+const updateUser = (user) => ({
+  type: UPDATE_USER,
+  user,
+});
 
 export const getSingleUser = (userId) => {
   return async (dispatch) => {
@@ -20,10 +26,25 @@ export const getSingleUser = (userId) => {
   };
 };
 
+export const updateUserWon = (id, points, currentLevel, currentGame, pidgeCoin) => {
+  return async (dispatch) => {
+      try {
+          const user = { points, currentLevel, currentGame, pidgeCoin }; 
+          const { data } = await axios.put(`/api/users/${id}`, user);
+          console.log("Thunk was dispatched!!")
+          dispatch(updateUser(data));
+      } catch (err) {
+          console.log("There was an error updating the user!", err)
+      }
+  }
+};
+
 export default function (state = {}, action) {
   switch (action.type) {
     case GET_SINGLE_USER:
       return action.data;
+    case UPDATE_USER:
+       return action.user;
     default:
       return state;
   }
