+++ server/api/users.js
@@ -19,7 +19,7 @@ router.get('/', async (req, res, next) => {
 router.get('/:userId', async (req, res, next) => {
   try {
     const user = await User.findByPk(req.params.userId, {
-      attributes: ['username', 'email', 'currentLevel', 'currentGame', 'points', 'streak', 'pidgeCoin'],
+      attributes: ['id', 'username', 'email', 'currentLevel', 'currentGame', 'points', 'streak', 'pidgeCoin'],
       include: [{ model: Pet }, { model: Badge }]
     });
     res.send(user)
